refactor: simplify timestamp logic and return types for different timestamp ranges
The ZTimeFormatter methods required many arguments, all of which were
the same from every user of the API. This PR simplifies a lot of these
calls and provides a clearer API for deciding exactly what sort of
timestamp you need using types for the different ranges and styles of
timestamp we may display.

This change was inspired by a translation issue in German where we need
to use a different string if we include a date vs including a time as a
format argument. Using types we know what sort of timestamp we have and
can now differentiate between the two.

To test, please check all timestamps in the application display correctly
and that the application does not crash on trying to render the timestamps.
In particular, please check on Android < 5.0 devices.